Fix a problem when 'text-transform' is used in navigation
When the 'text-transform' style is used, Selenium returns the user-visible transformed text, rather than the source text. When attempting to access the element via XPath, the transformed text was used, whereas the XPath was looking at the source text. In order to access that source text, the "textContent" property has to be used.

See https://github.com/SeleniumHQ/selenium/issues/3778 for more details on this issue.