Update cards iterator
the list(iter) should be valid, but I'm seeing exceptions trying to run
it in PDB debugging the test. Can't directly reproduce the error from
github actions for unit tests.

Seeing if this impacts the behavior in CI, since locally the test
appears to be running fine, and I can use the cards iterator to list the
many cards that are on the demo page.